Change class name form type + move files for better organization
